# lsv-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/region": {
      "get": {
        "operationId": "RegionController_getAllRegions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      },
      "post": {
        "operationId": "RegionController_createRegion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      }
    },
    "/region/{id}": {
      "get": {
        "operationId": "RegionController_getRegionById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      },
      "put": {
        "operationId": "RegionController_updateRegion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      },
      "delete": {
        "operationId": "RegionController_deleteRegion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      }
    },
...
    "/region/countries/search": {
      "get": {
        "operationId": "RegionController_searchCountries",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Region"
        ]
      }
    },
...
  },
  "info": {
    "title": "API LSV",
    "description": "Esquema OpenAPI de la API LSV",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateRegionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}